refactor: AvailableBeds로 변경
