--- src/main/java/com/curelingo/curelingo/egen/EgenApiClient.java
@@ -1,4 +1,0 @@
-package com.curelingo.curelingo.egen;
-
-public class EgenApiClient {
-}
