+++ src/main/java/com/curelingo/curelingo/egen/basic/HospitalInfoItem.java
@@ -1,0 +1,4 @@
+package com.curelingo.curelingo.egen.basic;
+
+public class HospitalInfoItem {
+}
