+++ src/main/java/com/curelingo/curelingo/egen/beds/AvailableBedsController.java
@@ -1,4 +1,4 @@
-package com.curelingo.curelingo.egen;
+package com.curelingo.curelingo.egen.beds;
 
 import com.curelingo.curelingo.egen.model.EgenResponse;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -8,19 +8,19 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RequestMapping("/api/egen")
 @RestController
-public class EgenController {
+public class AvailableBedsController {
 
-    private final EgenService egenService;
+    private final AvailableBedsService service;
 
-    public EgenController(EgenService egenService) {
-        this.egenService = egenService;
+    public AvailableBedsController(AvailableBedsService service) {
+        this.service = service;
     }
 
-    @GetMapping("")
-    public EgenResponse getEgen(
+    @GetMapping
+    public EgenResponse<AvailableBedsItem> getAvailableBeds(
             @RequestParam String stage1,
             @RequestParam String stage2
     ) {
-        return egenService.fetchEgenData(stage1, stage2);
+        return service.fetchAvailableBeds(stage1, stage2);
     }
 }
