+++ src/main/java/com/curelingo/curelingo/egen/model/EgenResponse.java
@@ -8,11 +8,11 @@ import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
 @JacksonXmlRootElement(localName = "response")
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class EgenResponse {
+public class EgenResponse<T> {
 
     @JacksonXmlProperty(localName = "header")
     public Header header;
 
     @JacksonXmlProperty(localName = "body")
-    public Body body;
+    public Body<T> body;
 }
